Add rendering system with `Renderer` component and asset handling improvements
- Introduced `Renderer` component with support for mesh and texture rendering.
- Integrated `mesh_holder` and `textures_holder` into `world_container` for dynamic asset management.
- Added support for game object rendering in `GameObject` and `world_container`.
- Improved logging macros for consistency and error handling.
- Refactored OpenGL initialization and cleanup processes in `r_window`.
- Enhanced `update` and `display` routines across components for seamless rendering workflow.
- Added new assets for testing, including models and textures.

+++ GameEngine-OpenGL-Freeglut-6/src/component/Camera.cpp
@@ -81,7 +81,7 @@ void Camera::update() {
 
     if (scrollDirection != 0) {
         float current_fov = rw->getFov();
-        float new_fov = current_fov + (scrollDirection * -2.0f);
+        float new_fov = current_fov + (static_cast<float>(scrollDirection) * -2.0f);
         rw->setFov(new_fov);
     }
